reorganize import/export of components
components are grouped as modules
