+++ src/components/about/index.js
@@ -1,0 +1,4 @@
+// components for the 'about' screen
+export { default as AboutListItem } from './aboutListItem';
+export { default as AboutListLink } from './aboutListLink';
+export { default as RedirectModal } from './redirectModal';
