+++ src/components/checkIn/index.js
@@ -1,0 +1,4 @@
+// components for the 'checkin' screen
+export { default as CheckInListView } from './checkInListView';
+export { default as CheckInTiles } from './checkInTiles';
+export { default as CheckInWelcomeText } from './welcomeText';
