+++ src/components/questionnaireModal/index.js
@@ -1,0 +1,4 @@
+// QuestionnaireModal component
+import QuestionnaireModal from './questionnaireModal';
+
+export default QuestionnaireModal;
