+++ src/components/shared/banner.jsx
