+++ src/components/shared/index.js
@@ -1,0 +1,8 @@
+// ScrollIndicatorWrapper component
+export { default as ScrollIndicatorWrapper } from './scrollIndicatorWrapper';
+
+// Spinner component for loading animation
+export { default as Spinner } from './spinner';
+
+// Banner Component
+export { default as Banner } from './banner';
