+++ src/components/shared/scrollIndicatorWrapper.jsx
