+++ src/screens/about/legalInformationScreen.jsx
@@ -8,9 +8,8 @@ import React, { PureComponent } from 'react';
 import { View, StyleSheet, Text } from 'react-native';
 
 import config from '../../config/configProvider';
-import Banner from '../../components/banner/banner';
+import { Banner, ScrollIndicatorWrapper } from '../../components/shared';
 import localization from '../../services/localization/localization';
-import ScrollIndicatorWrapper from '../../components/scrollIndicatorWrapper/scrollIndicatorWrapper';
 
 let localStyle;
